Memoize computation of style and move function to utils.
